move components into a new folder
